Refactor(utils): rename assertions.ts to api-utils.ts
